@@ -1,94 +1,85 @@
 {
+  "_args": [
+    [
+      "cordova-common@2.1.0",
+      "/Users/spindori/Documents/Cordova/cordova-ios"
+    ]
+  ],
+  "_from": "cordova-common@2.1.0",
+  "_id": "cordova-common@2.1.0",
+  "_inBundle": false,
+  "_integrity": "sha1-uzV+4bmCUDHtnbPFa1ku/pc9FkA=",
+  "_location": "/cordova-common",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "version",
+    "registry": true,
+    "raw": "cordova-common@2.1.0",
+    "name": "cordova-common",
+    "escapedName": "cordova-common",
+    "rawSpec": "2.1.0",
+    "saveSpec": null,
+    "fetchSpec": "2.1.0"
+  },
+  "_requiredBy": [
+    "/"
+  ],
+  "_resolved": "https://registry.npmjs.org/cordova-common/-/cordova-common-2.1.0.tgz",
+  "_spec": "2.1.0",
+  "_where": "/Users/spindori/Documents/Cordova/cordova-ios",
   "author": {
     "name": "Apache Software Foundation"
-  },
-  "name": "cordova-common",
-  "description": "Apache Cordova tools and platforms shared routines",
-  "license": "Apache-2.0",
-  "version": "1.1.1",
-  "repository": {
-    "type": "git",
-    "url": "git://git-wip-us.apache.org/repos/asf/cordova-common.git"
   },
   "bugs": {
     "url": "https://issues.apache.org/jira/browse/CB",
     "email": "dev@cordova.apache.org"
   },
-  "main": "cordova-common.js",
-  "engines": {
-    "node": ">=0.9.9"
-  },
-  "scripts": {
-    "test": "npm run jshint && npm run jasmine",
-    "jshint": "node node_modules/jshint/bin/jshint src && node node_modules/jshint/bin/jshint spec",
-    "jasmine": "node node_modules/jasmine-node/bin/jasmine-node --captureExceptions --color spec",
-    "cover": "node node_modules/istanbul/lib/cli.js cover --root src --print detail node_modules/jasmine-node/bin/jasmine-node -- spec"
-  },
-  "engineStrict": true,
+  "contributors": [],
   "dependencies": {
     "ansi": "^0.3.1",
     "bplist-parser": "^0.1.0",
     "cordova-registry-mapper": "^1.1.8",
-    "elementtree": "^0.1.6",
+    "elementtree": "0.1.6",
     "glob": "^5.0.13",
+    "minimatch": "^3.0.0",
     "osenv": "^0.1.3",
     "plist": "^1.2.0",
     "q": "^1.4.1",
     "semver": "^5.0.1",
-    "shelljs": "^0.5.1",
+    "shelljs": "^0.5.3",
     "underscore": "^1.8.3",
     "unorm": "^1.3.3"
   },
+  "description": "Apache Cordova tools and platforms shared routines",
   "devDependencies": {
-    "istanbul": "^0.3.17",
-    "jasmine-node": "^1.14.5",
-    "jshint": "^2.8.0"
+    "eslint": "^4.0.0",
+    "eslint-config-semistandard": "^11.0.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-plugin-import": "^2.3.0",
+    "eslint-plugin-node": "^5.0.0",
+    "eslint-plugin-promise": "^3.5.0",
+    "eslint-plugin-standard": "^3.0.1",
+    "istanbul": "^0.4.5",
+    "jasmine": "^2.5.2",
+    "promise-matchers": "^0.9.6",
+    "rewire": "^2.5.1"
   },
-  "contributors": [],
-  "_id": "cordova-common@1.1.1",
-  "_shasum": "4f74e182fe706b7a0ad626b0b3d303910af21d25",
-  "_resolved": "http://visustar.francelab.fr.ibm.com:8081/nexus/content/groups/mobile-npm-all/cordova-common/-/cordova-common-1.1.1.tgz",
-  "_from": "cordova-common@>=1.1.0 <2.0.0",
-  "_npmVersion": "3.5.3",
-  "_nodeVersion": "5.4.1",
-  "_npmUser": {
-    "name": "stevegill",
-    "email": "stevengill97@gmail.com"
+  "engines": {
+    "node": ">=4.0.0"
   },
-  "dist": {
-    "shasum": "4f74e182fe706b7a0ad626b0b3d303910af21d25",
-    "tarball": "http://visustar.francelab.fr.ibm.com:8081/nexus/content/groups/mobile-npm-all/cordova-common/-/cordova-common-1.1.1.tgz"
+  "homepage": "https://github.com/apache/cordova-lib#readme",
+  "license": "Apache-2.0",
+  "main": "cordova-common.js",
+  "name": "cordova-common",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/apache/cordova-lib.git"
   },
-  "maintainers": [
-    {
-      "name": "bowserj",
-      "email": "bowserj@apache.org"
-    },
-    {
-      "name": "kotikov.vladimir",
-      "email": "kotikov.vladimir@gmail.com"
-    },
-    {
-      "name": "purplecabbage",
-      "email": "purplecabbage@gmail.com"
-    },
-    {
-      "name": "shazron",
-      "email": "shazron@gmail.com"
-    },
-    {
-      "name": "stevegill",
-      "email": "stevengill97@gmail.com"
-    },
-    {
-      "name": "timbarham",
-      "email": "npmjs@barhams.info"
-    }
-  ],
-  "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/cordova-common-1.1.1.tgz_1458760117765_0.2676008273847401"
+  "scripts": {
+    "cover": "istanbul cover --root src --print detail jasmine",
+    "eslint": "eslint src && eslint spec",
+    "jasmine": "jasmine JASMINE_CONFIG_PATH=spec/support/jasmine.json",
+    "test": "npm run eslint && npm run jasmine"
   },
-  "directories": {},
-  "readme": "ERROR: No README data found!"
+  "version": "2.1.0"
 }
