@@ -1,81 +1,78 @@
 {
-  "name": "ios-sim",
-  "version": "5.0.8",
-  "preferGlobal": "true",
-  "description": "launch iOS apps into the iOS Simulator from the command line (Xcode 6.0+)",
-  "main": "ios-sim.js",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/phonegap/ios-sim.git"
-  },
-  "engines": {
-    "node": ">=0.10.0"
-  },
-  "keywords": [
-    "ios-sim",
-    "iOS Simulator"
+  "_args": [
+    [
+      "ios-sim@6.1.2",
+      "/Users/spindori/Documents/Cordova/cordova-ios"
+    ]
   ],
+  "_from": "ios-sim@6.1.2",
+  "_id": "ios-sim@6.1.2",
+  "_inBundle": false,
+  "_integrity": "sha512-kWSc7XJfYAGaQSh51UM4pDgXc77vqBS62oBA2tnXsHajk2+xl7Oyt613GYq64eMry2BM1F3tJLGFCJfpp1Ge+A==",
+  "_location": "/ios-sim",
+  "_phantomChildren": {
+    "abbrev": "1.1.1"
+  },
+  "_requested": {
+    "type": "version",
+    "registry": true,
+    "raw": "ios-sim@6.1.2",
+    "name": "ios-sim",
+    "escapedName": "ios-sim",
+    "rawSpec": "6.1.2",
+    "saveSpec": null,
+    "fetchSpec": "6.1.2"
+  },
+  "_requiredBy": [
+    "/"
+  ],
+  "_resolved": "https://registry.npmjs.org/ios-sim/-/ios-sim-6.1.2.tgz",
+  "_spec": "6.1.2",
+  "_where": "/Users/spindori/Documents/Cordova/cordova-ios",
+  "author": {
+    "name": "Shazron Abdullah"
+  },
   "bin": {
     "ios-sim": "./bin/ios-sim"
   },
   "bugs": {
     "url": "https://github.com/phonegap/ios-sim/issues"
   },
-  "author": {
-    "name": "Shazron Abdullah"
-  },
-  "license": "MIT",
   "dependencies": {
-    "plist": "^1.2.0",
-    "simctl": "^0.0.9",
+    "bplist-parser": "^0.0.6",
     "nopt": "1.0.9",
-    "bplist-parser": "^0.0.6"
+    "plist": "^1.2.0",
+    "simctl": "^1.1.1"
   },
+  "description": "launch iOS apps into the iOS Simulator from the command line (Xcode 8.0+)",
   "devDependencies": {
+    "jasmine": "~2.6.0",
+    "jscs": "^2.11.0",
     "jshint": "^2.9.1"
   },
-  "scripts": {
-    "test": "npm run jshint",
-    "jshint": "jshint src ./ios-sim.js"
+  "engines": {
+    "node": ">=4"
   },
-  "gitHead": "f0fc8160c262fe86ccaf8eeef1ff4ef22eb19b5c",
   "homepage": "https://github.com/phonegap/ios-sim#readme",
-  "_id": "ios-sim@5.0.8",
-  "_shasum": "20520bbcfdadb0b0cb08690b90969b6e60a9a796",
-  "_from": "ios-sim@>=5.0.7 <6.0.0",
-  "_npmVersion": "2.11.3",
-  "_nodeVersion": "0.12.7",
-  "_npmUser": {
-    "name": "shazron",
-    "email": "shazron@gmail.com"
-  },
-  "dist": {
-    "shasum": "20520bbcfdadb0b0cb08690b90969b6e60a9a796",
-    "tarball": "http://visustar.francelab.fr.ibm.com:8081/nexus/content/groups/mobile-npm-all/ios-sim/-/ios-sim-5.0.8.tgz"
-  },
-  "maintainers": [
-    {
-      "name": "macdonst",
-      "email": "simon.macdonald@gmail.com"
-    },
-    {
-      "name": "purplecabbage",
-      "email": "purplecabbage@gmail.com"
-    },
-    {
-      "name": "shazron",
-      "email": "shazron@gmail.com"
-    },
-    {
-      "name": "stevegill",
-      "email": "stevengill97@gmail.com"
-    }
+  "keywords": [
+    "ios-sim",
+    "iOS Simulator"
   ],
-  "_npmOperationalInternal": {
-    "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/ios-sim-5.0.8.tgz_1458757739978_0.3562360794749111"
+  "license": "MIT",
+  "main": "ios-sim.js",
+  "name": "ios-sim",
+  "preferGlobal": "true",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/phonegap/ios-sim.git"
   },
-  "directories": {},
-  "_resolved": "http://visustar.francelab.fr.ibm.com:8081/nexus/content/groups/mobile-npm-all/ios-sim/-/ios-sim-5.0.8.tgz",
-  "readme": "ERROR: No README data found!"
+  "scripts": {
+    "jasmine": "jasmine --config=spec/jasmine.json",
+    "jscs": "jscs src ./ios-sim.js",
+    "jshint": "jshint src ./ios-sim.js",
+    "postjshint": "npm run jscs",
+    "posttest": "npm run jshint",
+    "test": "npm run jasmine"
+  },
+  "version": "6.1.2"
 }
