@@ -1,82 +1,99 @@
 {
   "_args": [
     [
-      "pegjs@0.6.2",
-      "/Users/omefire/Projects/cordova-ios/node_modules/xcode"
+      "pegjs@0.10.0",
+      "/Users/spindori/Documents/Cordova/cordova-ios"
     ]
   ],
-  "_defaultsLoaded": true,
-  "_engineSupported": true,
-  "_from": "pegjs@0.6.2",
-  "_id": "pegjs@0.6.2",
-  "_inCache": true,
-  "_installable": true,
+  "_from": "pegjs@0.10.0",
+  "_id": "pegjs@0.10.0",
+  "_inBundle": false,
+  "_integrity": "sha1-z4uvrm7d/0tafvsYUmnqr0YQ3b0=",
   "_location": "/pegjs",
-  "_nodeVersion": "v0.4.8",
-  "_npmJsonOpts": {
-    "contributors": false,
-    "file": "/home/dmajda/.npm/pegjs/0.6.2/package/package.json",
-    "serverjs": false,
-    "wscript": false
-  },
-  "_npmVersion": "1.0.15",
   "_phantomChildren": {},
   "_requested": {
+    "type": "version",
+    "registry": true,
+    "raw": "pegjs@0.10.0",
     "name": "pegjs",
-    "raw": "pegjs@0.6.2",
-    "rawSpec": "0.6.2",
-    "scope": null,
-    "spec": "0.6.2",
-    "type": "version"
+    "escapedName": "pegjs",
+    "rawSpec": "0.10.0",
+    "saveSpec": null,
+    "fetchSpec": "0.10.0"
   },
   "_requiredBy": [
     "/xcode"
   ],
-  "_resolved": "https://registry.npmjs.org/pegjs/-/pegjs-0.6.2.tgz",
-  "_shasum": "74651f8a800e444db688e4eeae8edb65637a17a5",
-  "_shrinkwrap": null,
-  "_spec": "pegjs@0.6.2",
-  "_where": "/Users/omefire/Projects/cordova-ios/node_modules/xcode",
+  "_resolved": "https://registry.npmjs.org/pegjs/-/pegjs-0.10.0.tgz",
+  "_spec": "0.10.0",
+  "_where": "/Users/spindori/Documents/Cordova/cordova-ios",
   "author": {
-    "email": "david@majda.cz",
     "name": "David Majda",
+    "email": "david@majda.cz",
     "url": "http://majda.cz/"
   },
   "bin": {
     "pegjs": "bin/pegjs"
   },
   "bugs": {
-    "url": "https://github.com/dmajda/pegjs/issues"
+    "url": "https://github.com/pegjs/pegjs/issues"
   },
-  "dependencies": {},
   "description": "Parser generator for JavaScript",
   "devDependencies": {
-    "jake": ">= 0.1.10",
-    "uglify-js": ">= 0.0.5"
-  },
-  "directories": {},
-  "dist": {
-    "shasum": "74651f8a800e444db688e4eeae8edb65637a17a5",
-    "tarball": "http://registry.npmjs.org/pegjs/-/pegjs-0.6.2.tgz"
+    "browserify": "13.1.0",
+    "eslint": "2.13.1",
+    "http-server": "0.9.0",
+    "jasmine-node": "1.14.5",
+    "uglify-js": "2.7.0"
   },
   "engines": {
-    "node": ">= 0.4.4"
+    "node": ">=0.10"
   },
-  "homepage": "http://pegjs.majda.cz/",
-  "main": "lib/peg",
-  "maintainers": [
-    {
-      "email": "david@majda.cz",
-      "name": "dmajda"
-    }
+  "files": [
+    "CHANGELOG.md",
+    "LICENSE",
+    "README.md",
+    "VERSION",
+    "bin/pegjs",
+    "examples/arithmetics.pegjs",
+    "examples/css.pegjs",
+    "examples/javascript.pegjs",
+    "examples/json.pegjs",
+    "lib/compiler/asts.js",
+    "lib/compiler/index.js",
+    "lib/compiler/js.js",
+    "lib/compiler/opcodes.js",
+    "lib/compiler/passes/generate-bytecode.js",
+    "lib/compiler/passes/generate-js.js",
+    "lib/compiler/passes/remove-proxy-rules.js",
+    "lib/compiler/passes/report-duplicate-labels.js",
+    "lib/compiler/passes/report-duplicate-rules.js",
+    "lib/compiler/passes/report-infinite-recursion.js",
+    "lib/compiler/passes/report-infinite-repetition.js",
+    "lib/compiler/passes/report-undefined-rules.js",
+    "lib/compiler/visitor.js",
+    "lib/grammar-error.js",
+    "lib/parser.js",
+    "lib/peg.js",
+    "lib/utils/arrays.js",
+    "lib/utils/classes.js",
+    "lib/utils/objects.js",
+    "package.json"
   ],
+  "homepage": "http://pegjs.org/",
+  "keywords": [
+    "parser generator",
+    "PEG"
+  ],
+  "license": "MIT",
+  "main": "lib/peg",
   "name": "pegjs",
-  "optionalDependencies": {},
-  "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
-    "url": "git://github.com/dmajda/pegjs.git"
+    "url": "git+https://github.com/pegjs/pegjs.git"
   },
-  "scripts": {},
-  "version": "0.6.2"
+  "scripts": {
+    "test": "make lint && make spec"
+  },
+  "version": "0.10.0"
 }
