@@ -1,99 +1,68 @@
 {
   "_args": [
     [
-      "xmlbuilder@2.2.1",
-      "/Users/omefire/Projects/cordova-ios/node_modules/simple-plist/node_modules/plist"
+      "xmlbuilder@8.2.2",
+      "/Users/spindori/Documents/Cordova/cordova-ios"
     ]
   ],
-  "_from": "xmlbuilder@2.2.1",
-  "_id": "xmlbuilder@2.2.1",
-  "_inCache": true,
-  "_installable": true,
+  "_from": "xmlbuilder@8.2.2",
+  "_id": "xmlbuilder@8.2.2",
+  "_inBundle": false,
+  "_integrity": "sha1-aSSGc0ELS6QuGmE2VR0pIjNap3M=",
   "_location": "/simple-plist/xmlbuilder",
-  "_npmUser": {
-    "email": "oozcitak@gmail.com",
-    "name": "oozcitak"
-  },
-  "_npmVersion": "1.4.6",
   "_phantomChildren": {},
   "_requested": {
+    "type": "version",
+    "registry": true,
+    "raw": "xmlbuilder@8.2.2",
     "name": "xmlbuilder",
-    "raw": "xmlbuilder@2.2.1",
-    "rawSpec": "2.2.1",
-    "scope": null,
-    "spec": "2.2.1",
-    "type": "version"
+    "escapedName": "xmlbuilder",
+    "rawSpec": "8.2.2",
+    "saveSpec": null,
+    "fetchSpec": "8.2.2"
   },
   "_requiredBy": [
     "/simple-plist/plist"
   ],
-  "_resolved": "https://registry.npmjs.org/xmlbuilder/-/xmlbuilder-2.2.1.tgz",
-  "_shasum": "9326430f130d87435d4c4086643aa2926e105a32",
-  "_shrinkwrap": null,
-  "_spec": "xmlbuilder@2.2.1",
-  "_where": "/Users/omefire/Projects/cordova-ios/node_modules/simple-plist/node_modules/plist",
+  "_resolved": "https://registry.npmjs.org/xmlbuilder/-/xmlbuilder-8.2.2.tgz",
+  "_spec": "8.2.2",
+  "_where": "/Users/spindori/Documents/Cordova/cordova-ios",
   "author": {
-    "email": "oozcitak@gmail.com",
-    "name": "Ozgur Ozcitak"
+    "name": "Ozgur Ozcitak",
+    "email": "oozcitak@gmail.com"
   },
   "bugs": {
     "url": "http://github.com/oozcitak/xmlbuilder-js/issues"
   },
-  "dependencies": {
-    "lodash-node": "~2.4.1"
-  },
+  "contributors": [],
+  "dependencies": {},
   "description": "An XML builder for node.js",
   "devDependencies": {
-    "coffee-script": "~1.6.3",
-    "memwatch": "*",
-    "vows": "*"
-  },
-  "directories": {},
-  "dist": {
-    "shasum": "9326430f130d87435d4c4086643aa2926e105a32",
-    "tarball": "http://registry.npmjs.org/xmlbuilder/-/xmlbuilder-2.2.1.tgz"
+    "coffee-coverage": "*",
+    "coffee-script": "*",
+    "coveralls": "*",
+    "istanbul": "*",
+    "mocha": "*"
   },
   "engines": {
-    "node": "0.8.x || 0.10.x"
+    "node": ">=4.0"
   },
   "homepage": "http://github.com/oozcitak/xmlbuilder-js",
   "keywords": [
     "xml",
     "xmlbuilder"
   ],
-  "licenses": [
-    {
-      "type": "MIT",
-      "url": "http://opensource.org/licenses/mit-license.php"
-    }
-  ],
+  "license": "MIT",
   "main": "./lib/index",
-  "maintainers": [
-    {
-      "email": "oozcitak@gmail.com",
-      "name": "oozcitak"
-    }
-  ],
   "name": "xmlbuilder",
-  "optionalDependencies": {},
-  "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
     "url": "git://github.com/oozcitak/xmlbuilder-js.git"
   },
   "scripts": {
     "postpublish": "rm -rf lib",
-    "prepublish": "coffee -co lib/ src/*.coffee",
-    "test": "vows test/*"
+    "prepublish": "coffee -co lib src",
+    "test": "mocha && istanbul report text lcov"
   },
-  "version": "2.2.1",
-  "warnings": [
-    {
-      "code": "ENOTSUP",
-      "pkgid": "xmlbuilder@2.2.1",
-      "required": {
-        "node": "0.8.x || 0.10.x"
-      }
-    }
-  ]
+  "version": "8.2.2"
 }
