@@ -1,40 +1,35 @@
 {
   "_args": [
     [
-      "stream-buffers@~0.2.3",
-      "/Users/omefire/Projects/cordova-ios/node_modules/bplist-creator"
+      "stream-buffers@2.2.0",
+      "/Users/spindori/Documents/Cordova/cordova-ios"
     ]
   ],
-  "_from": "stream-buffers@>=0.2.3 <0.3.0",
-  "_id": "stream-buffers@0.2.6",
-  "_inCache": true,
-  "_installable": true,
+  "_from": "stream-buffers@2.2.0",
+  "_id": "stream-buffers@2.2.0",
+  "_inBundle": false,
+  "_integrity": "sha1-kdX1Ew0c75bc+n9yaUUYh0HQnuQ=",
   "_location": "/stream-buffers",
-  "_npmUser": {
-    "email": "me@samcday.com.au",
-    "name": "samcday"
-  },
-  "_npmVersion": "2.0.0-alpha-5",
   "_phantomChildren": {},
   "_requested": {
+    "type": "version",
+    "registry": true,
+    "raw": "stream-buffers@2.2.0",
     "name": "stream-buffers",
-    "raw": "stream-buffers@~0.2.3",
-    "rawSpec": "~0.2.3",
-    "scope": null,
-    "spec": ">=0.2.3 <0.3.0",
-    "type": "range"
+    "escapedName": "stream-buffers",
+    "rawSpec": "2.2.0",
+    "saveSpec": null,
+    "fetchSpec": "2.2.0"
   },
   "_requiredBy": [
     "/bplist-creator"
   ],
-  "_resolved": "https://registry.npmjs.org/stream-buffers/-/stream-buffers-0.2.6.tgz",
-  "_shasum": "181c08d5bb3690045f69401b9ae6a7a0cf3313fc",
-  "_shrinkwrap": null,
-  "_spec": "stream-buffers@~0.2.3",
-  "_where": "/Users/omefire/Projects/cordova-ios/node_modules/bplist-creator",
+  "_resolved": "https://registry.npmjs.org/stream-buffers/-/stream-buffers-2.2.0.tgz",
+  "_spec": "2.2.0",
+  "_where": "/Users/spindori/Documents/Cordova/cordova-ios",
   "author": {
-    "email": "me@samcday.com.au",
-    "name": "Sam Day"
+    "name": "Sam Day",
+    "email": "me@samcday.com.au"
   },
   "bugs": {
     "url": "https://github.com/samcday/node-stream-buffer/issues"
@@ -42,42 +37,27 @@
   "dependencies": {},
   "description": "Buffer-backed Streams for reading and writing.",
   "devDependencies": {
+    "istanbul": "~0.3.2",
     "vows": ">= 0.5.6"
   },
-  "directories": {},
-  "dist": {
-    "shasum": "181c08d5bb3690045f69401b9ae6a7a0cf3313fc",
-    "tarball": "http://registry.npmjs.org/stream-buffers/-/stream-buffers-0.2.6.tgz"
-  },
   "engines": {
-    "node": ">= 0.3.0"
+    "node": ">= 0.10.0"
   },
-  "gitHead": "b1dd62eb2e127d48ed7f1e814d5451d2b0dec144",
-  "homepage": "https://github.com/samcday/node-stream-buffer",
+  "homepage": "https://github.com/samcday/node-stream-buffer#readme",
   "keywords": [
     "memory streams",
     "streams",
     "buffer streams"
   ],
+  "license": "Unlicense",
   "main": "./lib/streambuffer.js",
-  "maintainers": [
-    {
-      "email": "sam.c.day@gmail.com",
-      "name": "samcday"
-    }
-  ],
   "name": "stream-buffers",
-  "optionalDependencies": {},
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org"
-  },
-  "readme": "ERROR: No README data found!",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/samcday/node-stream-buffer.git"
   },
   "scripts": {
-    "test": "vows --spec"
+    "test": "[ -n \"$NO_COVERAGE\" ] && vows --spec || istanbul cover vows -- --spec"
   },
-  "version": "0.2.6"
+  "version": "2.2.0"
 }
