@@ -32,10 +32,26 @@
 
 extern NSString* const CDVPageDidLoadNotification;
 extern NSString* const CDVPluginHandleOpenURLNotification;
+extern NSString* const CDVPluginHandleOpenURLWithAppSourceAndAnnotationNotification;
 extern NSString* const CDVPluginResetNotification;
-extern NSString* const CDVLocalNotification;
-extern NSString* const CDVRemoteNotification;
-extern NSString* const CDVRemoteNotificationError;
+extern NSString* const CDVViewWillAppearNotification;
+extern NSString* const CDVViewDidAppearNotification;
+extern NSString* const CDVViewWillDisappearNotification;
+extern NSString* const CDVViewDidDisappearNotification;
+extern NSString* const CDVViewWillLayoutSubviewsNotification;
+extern NSString* const CDVViewDidLayoutSubviewsNotification;
+extern NSString* const CDVViewWillTransitionToSizeNotification;
+
+/*
+ * The local and remote push notification functionality has been removed from the core in cordova-ios 4.x,
+ * but these constants have unfortunately have not been removed, but will be removed in 5.x.
+ * 
+ * To have the same functionality as 3.x, use a third-party plugin or the experimental
+ * https://github.com/apache/cordova-plugins/tree/master/notification-rebroadcast
+ */
+extern NSString* const CDVLocalNotification CDV_DEPRECATED(4.0, "Functionality removed in 4.0, constant will be removed in 5.0");
+extern NSString* const CDVRemoteNotification CDV_DEPRECATED(4.0, "Functionality removed in 4.0, constant will be removed in 5.0");
+extern NSString* const CDVRemoteNotificationError CDV_DEPRECATED(4.0, "Functionality removed in 4.0, constant will be removed in 5.0");
 
 @interface CDVPlugin : NSObject {}
 
@@ -50,6 +66,7 @@ extern NSString* const CDVRemoteNotificationError;
 - (void)pluginInitialize;
 
 - (void)handleOpenURL:(NSNotification*)notification;
+- (void)handleOpenURLWithApplicationSourceAndAnnotation:(NSNotification*)notification;
 - (void)onAppTerminate;
 - (void)onMemoryWarning;
 - (void)onReset;
@@ -61,7 +78,6 @@ extern NSString* const CDVRemoteNotificationError;
  - (void) onResume {}
  - (void) onOrientationWillChange {}
  - (void) onOrientationDidChange {}
- - (void)didReceiveLocalNotification:(NSNotification *)notification;
  */
 
 - (id)appDelegate;
