@@ -19,7 +19,7 @@
 
 //
 //  MainViewController.h
-//  HelloCordova
+//  Mobilot
 //
 //  Created by ___FULLUSERNAME___ on ___DATE___.
 //  Copyright ___ORGANIZATIONNAME___ ___YEAR___. All rights reserved.
