@@ -17,21 +17,26 @@
     under the License.
 */
 
-/*jslint node: true */
+/* jslint node: true */
 
 var fs = require('fs');
 var path = require('path');
-
+var unorm = require('unorm');
+var projectFile = require('./lib/projectFile');
+var check_reqs = require('./lib/check_reqs');
 var CordovaError = require('cordova-common').CordovaError;
 var CordovaLogger = require('cordova-common').CordovaLogger;
 var events = require('cordova-common').events;
+var PluginManager = require('cordova-common').PluginManager;
+var Q = require('q');
+var util = require('util');
 
-function setupEvents(externalEventEmitter) {
+function setupEvents (externalEventEmitter) {
     if (externalEventEmitter) {
         // This will make the platform internal events visible outside
         events.forwardEventsTo(externalEventEmitter);
     } else {
-        // There is no logger if external emitter is not present, 
+        // There is no logger if external emitter is not present,
         // so attach a console logger
         CordovaLogger.get().subscribe(events);
     }
@@ -48,10 +53,9 @@ function setupEvents(externalEventEmitter) {
  *   logging purposes. If no EventEmitter provided, all events will be logged to
  *   console
  */
-function Api(platform, platformRootDir, events) {
+function Api (platform, platformRootDir, events) {
     // 'platform' property is required as per PlatformApi spec
     this.platform = platform || 'ios';
-
     this.root = platformRootDir || path.resolve(__dirname, '..');
 
     setupEvents(events);
@@ -60,15 +64,15 @@ function Api(platform, platformRootDir, events) {
     var xcodeCordovaProj;
 
     try {
-        xcodeProjDir = fs.readdirSync(this.root).filter( function(e) { return e.match(/\.xcodeproj$/i); })[0];
+        xcodeProjDir = fs.readdirSync(this.root).filter(function (e) { return e.match(/\.xcodeproj$/i); })[0];
         if (!xcodeProjDir) {
             throw new CordovaError('The provided path "' + this.root + '" is not a Cordova iOS project.');
         }
 
-        var cordovaProjName = xcodeProjDir.substring(xcodeProjDir.lastIndexOf(path.sep)+1, xcodeProjDir.indexOf('.xcodeproj'));
+        var cordovaProjName = xcodeProjDir.substring(xcodeProjDir.lastIndexOf(path.sep) + 1, xcodeProjDir.indexOf('.xcodeproj'));
         xcodeCordovaProj = path.join(this.root, cordovaProjName);
-    } catch(e) {
-        throw new CordovaError('The provided path "'+this.root+'" is not a Cordova iOS project.');
+    } catch (e) {
+        throw new CordovaError('The provided path "' + this.root + '" is not a Cordova iOS project.');
     }
 
     this.locations = {
@@ -108,14 +112,25 @@ function Api(platform, platformRootDir, events) {
 Api.createPlatform = function (destination, config, options, events) {
     setupEvents(events);
 
-    return require('../../../lib/create')
-    .createProject(destination, config.packageName(), config.name(), options)
-    .then(function () {
-        // after platform is created we return Api instance based on new Api.js location
-        // This is required to correctly resolve paths in the future api calls
-        var PlatformApi = require(path.resolve(destination, 'cordova/Api'));
-        return new PlatformApi('ios', destination, events);
-    });
+    // CB-6992 it is necessary to normalize characters
+    // because node and shell scripts handles unicode symbols differently
+    // We need to normalize the name to NFD form since iOS uses NFD unicode form
+    var name = unorm.nfd(config.name());
+    var result;
+    try {
+        result = require('../../../lib/create')
+            .createProject(destination, config.packageName(), name, options)
+            .then(function () {
+                // after platform is created we return Api instance based on new Api.js location
+                // This is required to correctly resolve paths in the future api calls
+                var PlatformApi = require(path.resolve(destination, 'cordova/Api'));
+                return new PlatformApi('ios', destination, events);
+            });
+    } catch (e) {
+        events.emit('error', 'createPlatform is not callable from the iOS project API.');
+        throw (e);
+    }
+    return result;
 };
 
 /**
@@ -137,12 +152,19 @@ Api.createPlatform = function (destination, config, options, events) {
 Api.updatePlatform = function (destination, options, events) {
     setupEvents(events);
 
-    return require('../../../lib/create')
-    .updateProject(destination, options)
-    .then(function () {
-        var PlatformApi = require(path.resolve(destination, 'cordova/Api'));
-        return new PlatformApi('ios', destination, events);
-    });
+    var result;
+    try {
+        result = require('../../../lib/create')
+            .updateProject(destination, options)
+            .then(function () {
+                var PlatformApi = require(path.resolve(destination, 'cordova/Api'));
+                return new PlatformApi('ios', destination, events);
+            });
+    } catch (e) {
+        events.emit('error', 'updatePlatform is not callable from the iOS project API, you will need to do this manually.');
+        throw (e);
+    }
+    return result;
 };
 
 /**
@@ -198,8 +220,80 @@ Api.prototype.prepare = function (cordovaProject) {
  *   CordovaError instance.
  */
 Api.prototype.addPlugin = function (plugin, installOptions) {
-    var Plugman = require('./lib/plugman/Plugman');
-    return Plugman.get(this.locations).addPlugin(plugin, installOptions);
+    var xcodeproj = projectFile.parse(this.locations);
+    var self = this;
+
+    installOptions = installOptions || {};
+    installOptions.variables = installOptions.variables || {};
+    // Add PACKAGE_NAME variable into vars
+    if (!installOptions.variables.PACKAGE_NAME) {
+        installOptions.variables.PACKAGE_NAME = xcodeproj.getPackageName();
+    }
+
+    return PluginManager.get(self.platform, self.locations, xcodeproj)
+        .addPlugin(plugin, installOptions)
+        .then(function () {
+            var frameworkTags = plugin.getFrameworks(self.platform);
+            var frameworkPods = frameworkTags.filter(function (obj) {
+                return (obj.type === 'podspec');
+            });
+
+            return Q.resolve(frameworkPods);
+        })
+        .then(function (frameworkPods) {
+            if (!(frameworkPods.length)) {
+                return Q.resolve();
+            }
+
+            var project_dir = self.locations.root;
+            var project_name = self.locations.xcodeCordovaProj.split('/').pop();
+
+            var Podfile = require('./lib/Podfile').Podfile;
+            var PodsJson = require('./lib/PodsJson').PodsJson;
+
+            events.emit('verbose', 'Adding pods since the plugin contained <framework>(s) with type="podspec"');
+
+            var podsjsonFile = new PodsJson(path.join(project_dir, PodsJson.FILENAME));
+            var podfileFile = new Podfile(path.join(project_dir, Podfile.FILENAME), project_name);
+
+            frameworkPods.forEach(function (obj) {
+                var podJson = {
+                    name: obj.src,
+                    type: obj.type,
+                    spec: obj.spec
+                };
+
+                var val = podsjsonFile.get(podJson.name);
+                if (val) { // found
+                    if (podJson.spec !== val.spec) { // exists, different spec, print warning
+                        events.emit('warn', plugin.id + ' depends on ' + podJson.name + '@' + podJson.spec + ', which conflicts with another plugin. ' + podJson.name + '@' + val.spec + ' is already installed and was not overwritten.');
+                    }
+                    // increment count, but don't add in Podfile because it already exists
+                    podsjsonFile.increment(podJson.name);
+                } else { // not found, write new
+                    podJson.count = 1;
+                    podsjsonFile.setJson(podJson.name, podJson);
+                    // add to Podfile
+                    podfileFile.addSpec(podJson.name, podJson.spec);
+                }
+            });
+
+            // now that all the pods have been processed, write to pods.json
+            podsjsonFile.write();
+
+            // only write and pod install if the Podfile changed
+            if (podfileFile.isDirty()) {
+                podfileFile.write();
+                events.emit('verbose', 'Running `pod install` (to install plugins)');
+
+                return podfileFile.install(check_reqs.check_cocoapods);
+            } else {
+                events.emit('verbose', 'Podfile unchanged, skipping `pod install`');
+            }
+        })
+        // CB-11022 return non-falsy value to indicate
+        // that there is no need to run prepare after
+        .thenResolve(true);
 };
 
 /**
@@ -216,8 +310,69 @@ Api.prototype.addPlugin = function (plugin, installOptions) {
  *   CordovaError instance.
  */
 Api.prototype.removePlugin = function (plugin, uninstallOptions) {
-    var Plugman = require('./lib/plugman/Plugman');
-    return Plugman.get(this.locations).removePlugin(plugin, uninstallOptions);
+    var xcodeproj = projectFile.parse(this.locations);
+    var self = this;
+
+    return PluginManager.get(self.platform, self.locations, xcodeproj)
+        .removePlugin(plugin, uninstallOptions)
+        .then(function () {
+            var frameworkTags = plugin.getFrameworks(self.platform);
+            var frameworkPods = frameworkTags.filter(function (obj) {
+                return (obj.type === 'podspec');
+            });
+
+            return Q.resolve(frameworkPods);
+        })
+        .then(function (frameworkPods) {
+            if (!(frameworkPods.length)) {
+                return Q.resolve();
+            }
+
+            var project_dir = self.locations.root;
+            var project_name = self.locations.xcodeCordovaProj.split('/').pop();
+
+            var Podfile = require('./lib/Podfile').Podfile;
+            var PodsJson = require('./lib/PodsJson').PodsJson;
+
+            events.emit('verbose', 'Adding pods since the plugin contained <framework>(s) with type=\"podspec\"'); /* eslint no-useless-escape : 0 */
+
+            var podsjsonFile = new PodsJson(path.join(project_dir, PodsJson.FILENAME));
+            var podfileFile = new Podfile(path.join(project_dir, Podfile.FILENAME), project_name);
+
+            frameworkPods.forEach(function (obj) {
+                var podJson = {
+                    name: obj.src,
+                    type: obj.type,
+                    spec: obj.spec
+                };
+
+                var val = podsjsonFile.get(podJson.name);
+                if (val) { // found, decrement count
+                    podsjsonFile.decrement(podJson.name);
+                } else { // not found (perhaps a sync error)
+                    var message = util.format('plugin \"%s\" podspec \"%s\" does not seem to be in pods.json, nothing to remove. Will attempt to remove from Podfile.', plugin.id, podJson.name); /* eslint no-useless-escape : 0 */
+                    events.emit('verbose', message);
+                }
+
+                // always remove from the Podfile
+                podfileFile.removeSpec(podJson.name);
+            });
+
+            // now that all the pods have been processed, write to pods.json
+            podsjsonFile.write();
+
+            if (podfileFile.isDirty()) {
+                podfileFile.write();
+                events.emit('verbose', 'Running `pod install` (to uninstall pods)');
+
+                return podfileFile.install(check_reqs.check_cocoapods);
+            } else {
+                events.emit('verbose', 'Podfile unchanged, skipping `pod install`');
+            }
+        })
+        // CB-11022 return non-falsy value to indicate
+        // that there is no need to run prepare after
+        .thenResolve(true);
 };
 
 /**
@@ -254,10 +409,10 @@ Api.prototype.removePlugin = function (plugin, uninstallOptions) {
  */
 Api.prototype.build = function (buildOptions) {
     var self = this;
-    return require('./lib/check_reqs').run()
-    .then(function () {
-        return require('./lib/build').run.call(self, buildOptions);
-    });
+    return check_reqs.run()
+        .then(function () {
+            return require('./lib/build').run.call(self, buildOptions);
+        });
 };
 
 /**
@@ -272,12 +427,12 @@ Api.prototype.build = function (buildOptions) {
  * @return {Promise} A promise either fulfilled if package was built and ran
  *   successfully, or rejected with CordovaError.
  */
-Api.prototype.run = function(runOptions) {
+Api.prototype.run = function (runOptions) {
     var self = this;
-    return require('./lib/check_reqs').run()
-    .then(function () {
-        return require('./lib/run').run.call(self, runOptions);
-    });
+    return check_reqs.run()
+        .then(function () {
+            return require('./lib/run').run.call(self, runOptions);
+        });
 };
 
 /**
@@ -286,12 +441,15 @@ Api.prototype.run = function(runOptions) {
  * @return  {Promise}  Return a promise either fulfilled, or rejected with
  *   CordovaError.
  */
-Api.prototype.clean = function(cleanOptions) {
+Api.prototype.clean = function (cleanOptions) {
     var self = this;
-    return require('./lib/check_reqs').run()
-    .then(function () {
-        return require('./lib/clean').run.call(self, cleanOptions);
-    });
+    return check_reqs.run()
+        .then(function () {
+            return require('./lib/clean').run.call(self, cleanOptions);
+        })
+        .then(function () {
+            return require('./lib/prepare').clean.call(self, cleanOptions);
+        });
 };
 
 /**
@@ -302,8 +460,8 @@ Api.prototype.clean = function(cleanOptions) {
  * @return  {Promise<Requirement[]>}  Promise, resolved with set of Requirement
  *   objects for current platform.
  */
-Api.prototype.requirements = function() {
-    return require('./lib/check_reqs').check_all();
+Api.prototype.requirements = function () {
+    return check_reqs.check_all();
 };
 
 module.exports = Api;
