@@ -19,53 +19,51 @@
     under the License.
 */
 
-var child_process = require('child_process'),
-    Q = require('q');
+var child_process = require('child_process');
+var Q = require('q');
 
-exports.get_apple_ios_version = function() {
+exports.get_apple_ios_version = function () {
     var d = Q.defer();
-    child_process.exec('xcodebuild -showsdks', function(error, stdout, stderr) {
+    child_process.exec('xcodebuild -showsdks', function (error, stdout, stderr) {
         if (error) {
             d.reject(stderr);
-        }
-        else {
+        } else {
             d.resolve(stdout);
         }
     });
 
-    return d.promise.then(function(output) {
-        var regex = /[0-9]*\.[0-9]*/,
-            versions = [],
-            regexIOS = /^iOS \d+/;
+    return d.promise.then(function (output) {
+        var regex = /[0-9]*\.[0-9]*/;
+        var versions = [];
+        var regexIOS = /^iOS \d+/;
         output = output.split('\n');
         for (var i = 0; i < output.length; i++) {
             if (output[i].trim().match(regexIOS)) {
                 versions[versions.length] = parseFloat(output[i].match(regex)[0]);
-                }
+            }
         }
         versions.sort();
         console.log(versions[0]);
         return Q();
-    }, function(stderr) {
+    }, function (stderr) {
         return Q.reject(stderr);
     });
 };
 
-exports.get_apple_osx_version = function() {
+exports.get_apple_osx_version = function () {
     var d = Q.defer();
-    child_process.exec('xcodebuild -showsdks', function(error, stdout, stderr) {
+    child_process.exec('xcodebuild -showsdks', function (error, stdout, stderr) {
         if (error) {
             d.reject(stderr);
-        }
-        else {
+        } else {
             d.resolve(stdout);
         }
     });
 
-    return d.promise.then(function(output) {
-        var regex = /[0-9]*\.[0-9]*/,
-            versions = [],
-            regexOSX = /^OS X \d+/;
+    return d.promise.then(function (output) {
+        var regex = /[0-9]*\.[0-9]*/;
+        var versions = [];
+        var regexOSX = /^OS X \d+/;
         output = output.split('\n');
         for (var i = 0; i < output.length; i++) {
             if (output[i].trim().match(regexOSX)) {
@@ -75,14 +73,14 @@ exports.get_apple_osx_version = function() {
         versions.sort();
         console.log(versions[0]);
         return Q();
-    }, function(stderr) {
+    }, function (stderr) {
         return Q.reject(stderr);
     });
 };
 
-exports.get_apple_xcode_version = function() {
+exports.get_apple_xcode_version = function () {
     var d = Q.defer();
-    child_process.exec('xcodebuild -version', function(error, stdout, stderr) {
+    child_process.exec('xcodebuild -version', function (error, stdout, stderr) {
         var versionMatch = /Xcode (.*)/.exec(stdout);
         if (error || !versionMatch) {
             d.reject(stderr);
@@ -98,9 +96,26 @@ exports.get_apple_xcode_version = function() {
  * @return {Promise} Promise that either resolved with ios-deploy version
  *                           or rejected in case of error
  */
-exports.get_ios_deploy_version = function() {
+exports.get_ios_deploy_version = function () {
     var d = Q.defer();
-    child_process.exec('ios-deploy --version', function(error, stdout, stderr) {
+    child_process.exec('ios-deploy --version', function (error, stdout, stderr) {
+        if (error) {
+            d.reject(stderr);
+        } else {
+            d.resolve(stdout);
+        }
+    });
+    return d.promise;
+};
+
+/**
+ * Gets pod (CocoaPods) util version
+ * @return {Promise} Promise that either resolved with pod version
+ *                           or rejected in case of error
+ */
+exports.get_cocoapods_version = function () {
+    var d = Q.defer();
+    child_process.exec('pod --version', function (error, stdout, stderr) {
         if (error) {
             d.reject(stderr);
         } else {
@@ -115,9 +130,9 @@ exports.get_ios_deploy_version = function() {
  * @return {Promise} Promise that either resolved with ios-sim version
  *                           or rejected in case of error
  */
-exports.get_ios_sim_version = function() {
+exports.get_ios_sim_version = function () {
     var d = Q.defer();
-    child_process.exec('ios-sim --version', function(error, stdout, stderr) {
+    child_process.exec('ios-sim --version', function (error, stdout, stderr) {
         if (error) {
             d.reject(stderr);
         } else {
@@ -135,10 +150,11 @@ exports.get_ios_sim_version = function() {
  */
 exports.get_tool_version = function (toolName) {
     switch (toolName) {
-        case 'xcodebuild': return exports.get_apple_xcode_version();
-        case 'ios-sim': return exports.get_ios_sim_version();
-        case 'ios-deploy': return exports.get_ios_deploy_version();
-        default: return Q.reject(toolName + ' is not valid tool name. Valid names are: \'xcodebuild\', \'ios-sim\' and \'ios-deploy\'');
+    case 'xcodebuild': return exports.get_apple_xcode_version();
+    case 'ios-sim': return exports.get_ios_sim_version();
+    case 'ios-deploy': return exports.get_ios_deploy_version();
+    case 'pod': return exports.get_cocoapods_version();
+    default: return Q.reject(toolName + ' is not valid tool name. Valid names are: \'xcodebuild\', \'ios-sim\', \'ios-deploy\', and \'pod\'');
     }
 };
 
